Rename FilterSelect to Select and remove unneeded code
